requirements.txt
aiobotocore
aiohttp==3.7.4.post0
aioitertools==0.8.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
bidict==0.21.3
boto3
botocore
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
dnspython==1.16.0
docutils==0.15.2
fastapi==0.68.1
filelock==3.6.0
h11==0.12.0
huggingface-hub==0.5.1
idna==3.2
importlib-metadata==4.11.3
jmespath==0.10.0
joblib==1.1.0
motor==2.5.1
multidict==5.1.0
numpy==1.21.6
packaging==21.3
Pillow==8.3.2
pycodestyle==2.7.0
pydantic==1.8.2
pyfcm==1.5.4
PyJWT==2.1.0
pymongo==3.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.4
python-engineio==4.2.1
python-multipart==0.0.5
python-socketio==5.4.0
PyYAML==6.0
regex==2022.4.24
requests==2.26.0
s3transfer==0.5.0
sacremoses==0.0.49
sentencepiece==0.1.96
six==1.16.0
starlette==0.14.2
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
tqdm==4.64.0
transformers==4.18.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
websocket-client==1.2.1
websockets==10.0
wrapt==1.12.1
yarl==1.6.3
zipp==3.8.0
